fix(site): resolve circular dependency
