+++ site/src/pages/WorkspaceSettingsPage/WorkspaceSettingsLayout.tsx
@@ -2,38 +2,16 @@ import {
 	workspaceByOwnerAndName,
 	workspacePermissions,
 } from "api/queries/workspaces";
-import type { Workspace } from "api/typesGenerated";
 import { ErrorAlert } from "components/Alert/ErrorAlert";
 import { Loader } from "components/Loader/Loader";
 import { Margins } from "components/Margins/Margins";
 import { Stack } from "components/Stack/Stack";
-import type { WorkspacePermissions } from "modules/workspaces/permissions";
-import { createContext, type FC, Suspense, useContext } from "react";
+import { type FC, Suspense } from "react";
 import { useQuery } from "react-query";
 import { Outlet, useParams } from "react-router";
 import { pageTitle } from "utils/page";
 import { Sidebar } from "./Sidebar";
-
-type WorkspaceSettingsContext = {
-	owner: string;
-	workspace: Workspace;
-	permissions?: WorkspacePermissions;
-};
-
-const WorkspaceSettings = createContext<WorkspaceSettingsContext | undefined>(
-	undefined,
-);
-
-export function useWorkspaceSettings() {
-	const value = useContext(WorkspaceSettings);
-	if (!value) {
-		throw new Error(
-			"This hook can only be used from a workspace settings page",
-		);
-	}
-
-	return value;
-}
+import { WorkspaceSettings } from "./useWorkspaceSettings";
 
 export const WorkspaceSettingsLayout: FC = () => {
 	const params = useParams() as {
